linting fixes
