+++ tasks/change-granule-collection-s3/tests/test-index.js
@@ -14,7 +14,6 @@ const {
   putJsonS3Object,
   s3ObjectExists,
   promiseS3Upload,
-  parseS3Uri,
 } = require('@cumulus/aws-client/S3');
 const {
   randomId, validateOutput,
@@ -24,357 +23,18 @@ const {
   DeleteTopicCommand,
 } = require('@aws-sdk/client-sns');
 const { getDistributionBucketMapKey } = require('@cumulus/distribution-utils');
-const { isECHO10Filename, isUMMGFilename, metadataObjectFromCMRFile, granulesToCmrFileObjects, isCMRFile, getCMRCollectionId } = require('@cumulus/cmrjs/cmr-utils');
+const {
+  isECHO10Filename,
+  metadataObjectFromCMRFile,
+  granulesToCmrFileObjects,
+  isCMRFile,
+  getCMRCollectionId,
+} = require('@cumulus/cmrjs/cmr-utils');
 
 const { createSnsTopic } = require('@cumulus/aws-client/SNS');
 const { constructCollectionId } = require('../../../packages/message/Collections');
 const { changeGranuleCollectionS3, s3CopyNeeded, updateCMRData } = require('../dist/src');
-
-async function uploadFiles(files) {
-  await Promise.all(files.map((file) => {
-    let body;
-    if (isECHO10Filename(file)) {
-      body = fs.createReadStream('tests/data/meta.cmr.xml');
-    } else if (isUMMGFilename(file)) {
-      body = fs.createReadStream('tests/data/ummg-meta.cmr.json');
-    } else {
-      body = 'abc';
-    }
-    const parsedFile = parseS3Uri(file);
-    if (parsedFile.Bucket !== 'undefined' && parsedFile.Key !== 'undefindd') {
-      return promiseS3Upload({
-        params: {
-          ...parsedFile,
-          Body: body,
-        },
-      });
-    }
-    return null;
-  }));
-}
-function dummyGetCollection(collectionName, collectionVersion) {
-  return {
-    MOD11A1___001: {
-      files: [
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'protected',
-        },
-        {
-          regex: '^BROWSE\\.MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'BROWSE.MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf\\.met$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf.met',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.cmr\\.xml$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_2\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_1\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: 'public',
-          url_path: 'jpg/example2/',
-        },
-      ],
-      url_path: 'example2/{extractYear(cmrMetadata.Granule.Temporal.RangeDateTime.BeginningDateTime)}/',
-      name: 'MOD11A1',
-      granuleIdExtraction: '(MOD11A1\\.(.*))\\.hdf',
-      granuleId: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}$',
-      dataType: 'MOD11A1',
-      process: 'modis',
-      version: '001',
-      sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-      id: 'MOD11A2',
-    },
-    MOD11A1___002: {
-      files: [
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'protected',
-        },
-        {
-          regex: '^BROWSE\\.MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'BROWSE.MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf\\.met$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf.met',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.cmr\\.xml$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_2\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_1\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: 'public',
-          url_path: 'jpg/example2/',
-        },
-      ],
-      url_path: 'example2/{extractYear(cmrMetadata.Granule.Temporal.RangeDateTime.BeginningDateTime)}/',
-      name: 'MOD11A1',
-      granuleIdExtraction: '(MOD11A1\\.(.*))\\.hdf',
-      granuleId: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}$',
-      dataType: 'MOD11A1',
-      process: 'modis',
-      version: '002',
-      sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-      id: 'MOD11A1',
-    },
-    MOD11A1UMMG___001: {
-      files: [
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'protected',
-        },
-        {
-          regex: '^BROWSE\\.MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'BROWSE.MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf\\.met$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf.met',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.ummg\\.cmr\\.json$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.iso.xml',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_2\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_1\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: 'public',
-          url_path: 'jpg/example2/',
-        },
-      ],
-      url_path: 'example2/{extractYear(cmrMetadata.TemporalExtent.RangeDateTime.BeginningDateTime)}/',
-      name: 'MOD11A1',
-      granuleIdExtraction: '(MOD11A1\\.(.*))\\.hdf',
-      granuleId: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}$',
-      dataType: 'MOD11A1',
-      process: 'modis',
-      version: '001',
-      sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-      id: 'MOD11A1',
-    },
-    MOD11A1UMMG___002: {
-      files: [
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'protected',
-        },
-        {
-          regex: '^BROWSE\\.MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'BROWSE.MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf\\.met$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf.met',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.ummg\\.cmr\\.json$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.iso.xml',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_2\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_1\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: 'public',
-          url_path: 'jpg/example2/',
-        },
-      ],
-      url_path: 'example2/{extractYear(cmrMetadata.TemporalExtent.RangeDateTime.BeginningDateTime)}/',
-      name: 'MOD11A1',
-      granuleIdExtraction: '(MOD11A1\\.(.*))\\.hdf',
-      granuleId: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}$',
-      dataType: 'MOD11A1',
-      process: 'modis',
-      version: '002',
-      sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-      id: 'MOD11A1',
-    },
-    MOD11ANOMOVE___001: {
-      files: [
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.hdf$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: 'protected',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_1\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: 'private',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}_2\\.jpg$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: 'public',
-        },
-        {
-          regex: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}\\.cmr\\.xml$',
-          sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          bucket: 'protected',
-        },
-      ],
-      url_path: 'file-staging/subdir/',
-      name: 'MOD11A1',
-      granuleIdExtraction: '(MOD11A1\\.(.*))\\.hdf',
-      granuleId: '^MOD11A1\\.A[\\d]{7}\\.[\\S]{6}\\.006.[\\d]{13}$',
-      dataType: 'MOD11A1',
-      process: 'modis',
-      version: '001',
-      sampleFileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-      id: 'MOD11A1',
-    },
-  }[constructCollectionId(collectionName, collectionVersion)];
-}
-
-function dummyGetGranule(granuleId, t) {
-  const granuleSet = {
-    base_xml_granule: {
-      status: 'completed',
-      collectionId: 'MOD11A1___006',
-      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090724',
-      files: [
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          bucket: t.context.protectedBucket,
-          type: 'data',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: t.context.privateBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: t.context.publicBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          bucket: t.context.protectedBucket,
-          type: 'metadata',
-        },
-      ],
-    },
-    base_umm_granule: {
-      status: 'completed',
-      collectionId: 'MOD11A1___006',
-      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090725',
-      files: [
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725.hdf',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725.hdf',
-          bucket: t.context.protectedBucket,
-          type: 'data',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725_1.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725_1.jpg',
-          bucket: t.context.privateBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725_2.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725_2.jpg',
-          bucket: t.context.publicBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725.ummg.cmr.json',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725.ummg.cmr.json',
-          bucket: t.context.protectedBucket,
-          type: 'metadata',
-        },
-      ],
-    },
-    bad_granule: {
-      status: 'completed',
-      collectionId: 'MOD11A1___006',
-      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090724',
-      files: [
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
-          type: 'data',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
-          bucket: t.context.privateBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
-          bucket: t.context.publicBucket,
-          type: 'browse',
-        },
-        {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
-          bucket: t.context.protectedBucket,
-          type: 'metadata',
-        },
-      ],
-    },
-    empty_xml_granule: {
-      status: 'completed',
-      collectionId: 'MOD11A1___006',
-      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090724',
-      files: [],
-    },
-  };
-  range(200).forEach((i) => {
-    const baseGranuleString = JSON.stringify(
-      granuleSet.base_xml_granule
-    ).replaceAll('90724', `${('0000' + i).slice(-5)}`);
-    granuleSet[`xml_granule${i}`] = JSON.parse(baseGranuleString);
-  });
-  return granuleSet[granuleId];
-}
+const { dummyGetCollection, dummyGetGranule, uploadFiles } = require('./_helpers');
 
 function granulesToFileURIs(granuleIds, t) {
   const granules = granuleIds.map((granuleId) => dummyGetGranule(granuleId, t));
